optimized edit profile feature & updates to app.js
edit profile optimized- hang time reduced to 3 secs.

app.js- NEW FEATURE THAT ALLOWS MULTIPLE PORT NUMBERS. The default is still 3000, but if you put a port number (i.e. npm start 8888), routes will be running on that port number. This is so that we can test with multiple websites simultaneously.